requirements.txt
# Core dependencies
# Note: Requires Python 3.12

# Google Agent Development Kit (ADK)
google-adk>=1.2.0
# Note: ADK is installed from GitHub: pip install git+https://github.com/google/adk-python.git
# If you don't have the ADK directory, clone from: https://github.com/google/adk-python

# Google Cloud dependencies
google-cloud-pubsub>=2.18.0
google-cloud-logging>=3.8.0
google-cloud-storage>=2.10.0
google-cloud-compute>=1.14.0
google-cloud-aiplatform>=1.36.0  # For Vertex AI / Gemini
google-cloud-bigquery>=3.13.0
google-cloud-firestore>=2.13.0
google-cloud-secret-manager>=2.17.0
google-cloud-functions>=1.14.0
google-cloud-monitoring>=2.16.0
google-cloud-kms>=2.19.0
google-cloud-iam>=2.12.0
# google-cloud-iam-credentials  # Not available as a separate package
google-auth>=2.23.0
google-api-core>=2.14.0

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Async support
aiohttp>=3.8.0
# asyncio is part of Python standard library, no need to install

# Data handling
pyyaml>=6.0
python-json-logger>=2.0.7
jsonschema>=4.19.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.6.0  # For better CLI output

# Testing dependencies (development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-timeout>=2.1.0
httpx>=0.24.0  # For FastAPI test client

# Development tools
black>=23.9.0
flake8>=6.1.0
mypy>=1.5.0
pylint>=3.0.0
isort>=5.12.0
bandit>=1.7.5  # Security linter
radon>=6.0.1   # Code complexity checker

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0

# Type stubs
types-pyyaml>=6.0.0
types-requests>=2.31.0

# Monitoring and observability
prometheus-client>=0.18.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-instrumentation-requests>=0.41b0
opentelemetry-instrumentation-grpc>=0.41b0
opentelemetry-exporter-gcp-trace>=1.6.0

# Communication services
twilio>=8.0.0  # For SMS notifications
slack-sdk>=3.23.0  # For Slack notifications

# Security
cryptography>=41.0.0
pyjwt>=2.8.0

# Rate limiting
slowapi>=0.1.9

# System monitoring
psutil>=5.9.0

# Background tasks
celery>=5.3.0
redis>=5.0.0

# Database (if needed later)
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.28.0

# Cost optimization and data analysis
google-cloud-billing-budgets>=1.12.0
pandas>=2.1.0
numpy>=1.25.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
matplotlib>=3.7.0
seaborn>=0.12.0
schedule>=1.2.0
